requirements.txt
# Essential packages for FigVarietyRAGChat

# Core LangChain components
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface

# Vector store
faiss-cpu

# LLM and Embeddings support
google-generativeai
sentence-transformers

# Utilities
python-dotenv